added admin edit post, delete post

+++ controllers/avatarUser.js
@@ -27,11 +27,46 @@ exports.getAvatarUser = (req, res, next) => {
 }
 
 exports.deleteAvatar = (req, res, next) => {
-    AvatarModel.findOne({ _id: req.params.id }).then((avatar) => {
-        AvatarModel.deleteOne({ _id: req.params.id })
-            .then(() => {
-                res.status(200).json({ message: "Avatar supprimé" })
-            })
-            .catch((err) => console.log(err))
-    })
+    AvatarModel.findOne({ _id: req.params.id })
+        .then((avatar) => {
+            if (avatar.userId != req.body.userId) {
+                res.status(401).json({ message: "Not authorized to do this." })
+            } else {
+                AvatarModel.deleteOne({ _id: req.params.id })
+                    .then(() => {
+                        res.status(200).json({ message: "Avatar supprimé" })
+                    })
+                    .catch((err) => console.log(err))
+            }
+        })
+        .catch((err) => console.log(err))
+}
+
+exports.editAvatar = (req, res, next) => {
+    console.log(req.body)
+    const obj = req.file
+        ? {
+              ...JSON.parse(req.body),
+              imageUrl: `${req.protocol}://${req.get("host")}/images/avatar/${
+                  req.file.filename
+              }`,
+          }
+        : { ...req.body }
+    delete obj._userId
+    AvatarModel.findOne({ _id: req.params.id })
+        .then((avatar) => {
+            if (avatar.userId != req.body.userId) {
+                res.status(401).json({ message: "Not authorized to do this." })
+            } else {
+                AvatarModel.updateOne(
+                    { _id: req.params.id },
+                    { ...req.body, _id: req.params.id }
+                )
+                    .then(() => {
+                        res.status(200).json({ message: "Avatar updated" })
+                    })
+                    .catch((error) => res.status(400).json({ error }))
+            }
+        })
+        .catch((error) => res.status(401).json({ error }))
 }
